\name{DynStateInput-class}
\docType{class}
\alias{DynStateInput-class}

\title{Class DynStateInput}
\description{ A class to represent expected catchrates with mean and CV  }
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("FLBiol", ...)}.
}
\section{Slots}{
  \describe{
   
    \item{\code{CatchMean}:}{Object of class \code{"FLQuant"}, with the average catch rate  }
    \item{\code{CatchSigma}:}{Object of class \code{"FLQuant"}, with the sigma of the ctach rate, assuming a normal distribution }
  }
}
\section{Methods}{

}
\author{Jan Jaap Poos}
\examples{
}
\keyword{classes}
